+++ src/main/java/io/github/ryuu/adventurecraft/gui/GuiSlider2.java
@@ -1,5 +1,7 @@
 package io.github.ryuu.adventurecraft.gui;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.widgets.Button;
 import org.lwjgl.opengl.GL11;
@@ -34,7 +36,7 @@ public class GuiSlider2 extends Button {
                 this.sliderValue = 1.0f;
             }
         }
-        GL11.glColor4f(1.0f, 1.0f, 1.0f, 1.0f);
+        GL11.glColor4f((float) 1.0f, (float) 1.0f, (float) 1.0f, (float) 1.0f);
         this.blit(this.x + (int) (this.sliderValue * (float) (this.width - 8)), this.y, 0, 66, 4, 20);
         this.blit(this.x + (int) (this.sliderValue * (float) (this.width - 8)) + 4, this.y, 196, 66, 4, 20);
     }
