+++ src/main/java/io/github/ryuu/adventurecraft/scripting/ScriptUISprite.java
@@ -1,7 +1,13 @@
 package io.github.ryuu.adventurecraft.scripting;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.render.Tessellator;
+import net.minecraft.client.render.TextRenderer;
+import net.minecraft.client.texture.TextureManager;
+import net.minecraft.script.ScriptUIContainer;
+import net.minecraft.script.UIElement;
 import org.lwjgl.opengl.GL11;
 
 public class ScriptUISprite extends UIElement {
@@ -52,7 +58,7 @@ public class ScriptUISprite extends UIElement {
         } else {
             renderEngine.bindTexture(renderEngine.getTextureId(this.texture));
         }
-        GL11.glColor4f(this.red, this.green, this.blue, this.alpha);
+        GL11.glColor4f((float) this.red, (float) this.green, (float) this.blue, (float) this.alpha);
         float x = this.getXAtTime(partialTickTime);
         float y = this.getYAtTime(partialTickTime);
         float f = 1.0f / this.imageWidth;
